Update vendored gopkg.in/yaml.v2.
