+++ Gopkg.toml
@@ -198,3 +198,7 @@ required = [
 [[constraint]]
   name = "go.etcd.io/bbolt"
   version = "1.3.0"
+
+[[constraint]]
+  name = "gopkg.in/yaml.v2"
+  version = "2.1.1"
