+++ cli/client/config/config.go
@@ -5,21 +5,30 @@ import (
 )
 
 const (
-	// DefaultEdition represents the default Sensu edition
+	// DefaultEdition is the default Sensu edition
 	DefaultEdition = types.CoreEdition
-	// DefaultFormat represents the default format output when displaying objects
-	DefaultFormat = FormatTabular
+
 	// DefaultNamespace represents the default namespace
 	DefaultNamespace = "default"
-	// FormatTabular represents the string for tabular format
+
+	// DefaultFormat is the default format output for printers.
+	DefaultFormat = FormatTabular
+
+	// FormatTabular indicates tabular format for printers.
 	FormatTabular = "tabular"
-	// FormatJSON represents the string for JSON format
+
+	// FormatJSON indicates JSON format for printers.
 	FormatJSON = "json"
-	// FormatWrappedJSON represents the string for wrapped JSON format
+
+	// FormatWrappedJSON indicates wrapped JSON format for printers.
 	FormatWrappedJSON = "wrapped-json"
+
+	// FormatYAML indicates YAML format for printers. It has the same layout
+	// as wrapped JSON.
+	FormatYAML = "yaml"
 )
 
-// Config represents an abstracted configuration
+// Config is an abstract configuration
 type Config interface {
 	Read
 	Write
